Disable all overrides of formatoptions
At first I just wanted to remove the "o" option from it which makes
pressing "o" in Normal mode on a commented line start the new line as a
commented line, which is almost never what I want. But after thinking
about it, I really don't want the options that are available in
formatoptions to change per filetype... I want them to always be the
same. This is done with an autocommand on BufEnter.